Still unable to get the counter to work. Did not get to difficulty level, nor able to make the game work.
